Remove customization from appdirs
