+++ conda/gateways/anaconda_client.py
@@ -10,7 +10,7 @@ from os.path import isdir, isfile, join
 from stat import S_IREAD, S_IWRITE
 
 from .disk.delete import rm_rf
-from .._vendor.appdirs import AppDirs, EnvAppDirs
+from .._vendor.appdirs import AppDirs
 from ..common.compat import scandir
 from ..common.url import quote_plus, unquote_plus
 
@@ -20,6 +20,29 @@ log = getLogger(__name__)
 def replace_first_api_with_conda(url):
     # replace first occurrence of 'api' with 'conda' in url
     return re.sub(r'([./])api([./]|$)', r'\1conda\2', url, count=1)
+
+
+class EnvAppDirs:
+    def __init__(self, appname, appauthor, root_path):
+        self.appname = appname
+        self.appauthor = appauthor
+        self.root_path = root_path
+
+    @property
+    def user_data_dir(self):
+        return join(self.root_path, "data")
+
+    @property
+    def site_data_dir(self):
+        return join(self.root_path, "data")
+
+    @property
+    def user_cache_dir(self):
+        return join(self.root_path, "cache")
+
+    @property
+    def user_log_dir(self):
+        return join(self.root_path, "log")
 
 
 def _get_binstar_token_directory():
